V1 Views refactorized
