+++ gateway/api/v1/views/__init__.py
@@ -1,0 +1,10 @@
+"""
+Django Rest framework ViewSets for api application:
+
+Specification for V1 version of the API.
+"""
+
+from .catalog import CatalogViewSet
+from .files import FilesViewSet
+from .jobs import JobViewSet
+from .programs import ProgramViewSet
