+++ gateway/api/v1/views/files.py
@@ -1,0 +1,16 @@
+"""
+Files view api for V1.
+"""
+
+from rest_framework import permissions
+
+from api.permissions import IsOwner
+from api import views
+
+
+class FilesViewSet(views.FilesViewSet):
+    """
+    Files view set.
+    """
+
+    permission_classes = [permissions.IsAuthenticated, IsOwner]
