+++ gateway/api/v1/views/jobs.py
@@ -1,0 +1,49 @@
+"""
+Jobs view api for V1.
+"""
+
+from drf_yasg.utils import swagger_auto_schema
+from rest_framework import permissions, status
+from rest_framework.decorators import action
+from rest_framework.pagination import LimitOffsetPagination
+
+from api import views
+from api.permissions import IsOwner
+from api.v1 import serializers as v1_serializers
+
+
+class JobViewSet(views.JobViewSet):
+    """
+    Job view set first version. Use JobSerializer V1.
+    """
+
+    serializer_class = v1_serializers.JobSerializer
+    pagination_class = LimitOffsetPagination
+    permission_classes = [permissions.IsAuthenticated, IsOwner]
+
+    def get_serializer_class(self):
+        return v1_serializers.JobSerializer
+
+    @swagger_auto_schema(
+        operation_description="Get author Job",
+        responses={status.HTTP_200_OK: v1_serializers.JobSerializer(many=False)},
+    )
+    def retrieve(self, request, pk=None):
+        return super().retrieve(request, pk)
+
+    @swagger_auto_schema(
+        operation_description="List author Jobs",
+        responses={status.HTTP_200_OK: v1_serializers.JobSerializer(many=True)},
+    )
+    def list(self, request):
+        return super().list(request)
+
+    @swagger_auto_schema(
+        operation_description="Save the result of a job",
+        responses={status.HTTP_200_OK: v1_serializers.JobSerializer(many=False)},
+    )
+    @action(methods=["POST"], detail=True)
+    def result(self, request, pk=None):
+        return super().result(request, pk)
+
+    ### We are not returning serializers in the rest of the end-points
