+++ gateway/api/v1/views/programs.py
@@ -1,0 +1,66 @@
+"""
+Programs view api for V1.
+"""
+
+# pylint: disable=duplicate-code
+from drf_yasg.utils import swagger_auto_schema
+from rest_framework import permissions, status
+from rest_framework.decorators import action
+
+from api import views
+from api.v1 import serializers as v1_serializers
+
+
+class ProgramViewSet(views.ProgramViewSet):
+    """
+    Quantum function view set first version. Use ProgramSerializer V1.
+    """
+
+    serializer_class = v1_serializers.ProgramSerializer
+    pagination_class = None
+    permission_classes = [permissions.IsAuthenticated]
+
+    @staticmethod
+    def get_serializer_job_config(*args, **kwargs):
+        return v1_serializers.JobConfigSerializer(*args, **kwargs)
+
+    @staticmethod
+    def get_serializer_upload_program(*args, **kwargs):
+        return v1_serializers.UploadProgramSerializer(*args, **kwargs)
+
+    @staticmethod
+    def get_serializer_run_program(*args, **kwargs):
+        return v1_serializers.RunProgramSerializer(*args, **kwargs)
+
+    @staticmethod
+    def get_serializer_run_job(*args, **kwargs):
+        return v1_serializers.RunJobSerializer(*args, **kwargs)
+
+    @staticmethod
+    def get_serializer_job(*args, **kwargs):
+        return v1_serializers.JobSerializer(*args, **kwargs)
+
+    @swagger_auto_schema(
+        operation_description="List author Qiskit Functions",
+        responses={status.HTTP_200_OK: v1_serializers.ProgramSerializer(many=True)},
+    )
+    def list(self, request):
+        return super().list(request)
+
+    @swagger_auto_schema(
+        operation_description="Upload a Qiskit Function",
+        request_body=v1_serializers.UploadProgramSerializer,
+        responses={status.HTTP_200_OK: v1_serializers.UploadProgramSerializer},
+    )
+    @action(methods=["POST"], detail=False)
+    def upload(self, request):
+        return super().upload(request)
+
+    @swagger_auto_schema(
+        operation_description="Run an existing Qiskit Function",
+        request_body=v1_serializers.RunProgramSerializer,
+        responses={status.HTTP_200_OK: v1_serializers.RunJobSerializer},
+    )
+    @action(methods=["POST"], detail=False)
+    def run(self, request):
+        return super().run(request)
